+++ CHANGELOG.md
@@ -2,5 +2,6 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
+* Scaffolded the provider based on v2.6.17 of the Artifactory Terraform Provider
 
 ---